docs: prose layout all docs, docs for getters
