+++ libs/playground/src/app/docs/block/page.mdx
@@ -46,7 +46,7 @@ The UI above is made out of blocks like this:
 
 <Prose className="mt-6">
 
-## `useBlock`
+## `useBlock` hook
 
 `useBlock` is a utility hook that allows for access to similar features as `Block`.
