+++ libs/playground/src/app/docs/chat/page.mdx
