+++ libs/playground/src/app/docs/image/UseImageGeneration.tsx
